A bit cleaner code. Add tool change and spindle speed parsing centralized. Carry forward the feed rate, as well as x/y/z

+++ lib/aid3/sim/tooling/cnc.go
@@ -81,7 +81,7 @@ type Cnc interface {
 	FastFeedRate() float64
 	FeedMode(mode int)
 	SpindleSpeed(speed int64)
-	ToolChangeTo(tool int)
+	ToolChangeTo(tool int64)
 	SelectPlane(plane int)
 	Plane() int
 	Reset()
